@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\Action;
 
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\PageModel;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;
@@ -16,12 +16,8 @@ final class FacebookLintAction
 {
     private const FACEBOOK_LINT_URL = 'https://developers.facebook.com/tools/debug/?q=';
 
-    /** @var ContaoFrameworkInterface */
-    private $framework;
-
-    public function __construct(ContaoFrameworkInterface $framework)
+    public function __construct(private ContaoFramework $framework)
     {
-        $this->framework = $framework;
     }
 
     public function __invoke(int $pageId): Response
