@@ -4,28 +4,26 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\Data\Extractor;
 
-use Contao\Controller;
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\Framework\ContaoFramework;
+use Contao\CoreBundle\InsertTag\InsertTagParser;
+use Contao\CoreBundle\Routing\ResponseContext\HtmlHeadBag\HtmlHeadBag;
+use Contao\CoreBundle\Routing\ResponseContext\ResponseContextAccessor;
 use Contao\FilesModel;
 use Hofff\Contao\SocialTags\Data\Extractor;
 use Symfony\Component\HttpFoundation\RequestStack;
 
+use function is_file;
+
+/** @SuppressWarnings(PHPMD.LongVariable) */
 abstract class AbstractExtractor implements Extractor
 {
-    /** @var RequestStack */
-    protected $requestStack;
-
-    /** @var ContaoFrameworkInterface */
-    protected $framework;
-
-    /** @var string */
-    protected $projectDir;
-
-    public function __construct(ContaoFrameworkInterface $framework, RequestStack $requestStack, string $projectDir)
-    {
-        $this->framework    = $framework;
-        $this->projectDir   = $projectDir;
-        $this->requestStack = $requestStack;
+    public function __construct(
+        protected ContaoFramework $framework,
+        protected RequestStack $requestStack,
+        protected ResponseContextAccessor $responseContextAccessor,
+        protected InsertTagParser $insertTagParser,
+        protected string $projectDir,
+    ) {
     }
 
     protected function getBaseUrl(): string
@@ -36,7 +34,7 @@ abstract class AbstractExtractor implements Extractor
             return $baseUrl;
         }
 
-        $request = $this->requestStack->getMasterRequest();
+        $request = $this->requestStack->getMainRequest();
         if (! $request) {
             return '';
         }
@@ -48,7 +46,7 @@ abstract class AbstractExtractor implements Extractor
 
     protected function getRequestUri(): string
     {
-        $request = $this->requestStack->getMasterRequest();
+        $request = $this->requestStack->getMainRequest();
         if (! $request) {
             return '/';
         }
@@ -56,20 +54,66 @@ abstract class AbstractExtractor implements Extractor
         return $request->getRequestUri();
     }
 
-    protected function getFileModel(string $uuid): ?FilesModel
+    /**
+     * Retrieves an image from the reference or fallback object.
+     *
+     * Resolving an image is done by following steps:
+     *  - Check if reference object enables hofff_st and has an image configured by $key
+     *  - Check if reference object enables addImage and has an image configured by singleSRC
+     *  - Check if fallback object provides an image by $key
+     */
+    protected function getImage(
+        string $key,
+        object $reference,
+        object|null $fallback = null,
+    ): FilesModel|null {
+        if ($reference->hofff_st && $reference->{$key}) {
+            $image = $reference->{$key};
+        } elseif ($reference->addImage && $reference->singleSRC) {
+            $image = $reference->singleSRC;
+        } elseif ($fallback && $fallback->{$key}) {
+            $image = $fallback->{$key};
+        } else {
+            return null;
+        }
+
+        return $this->getFileModel($image);
+    }
+
+    protected function getFileUrl(FilesModel|null $file = null): string|null
+    {
+        if ($file && is_file($this->projectDir . '/' . $file->path)) {
+            return $this->getBaseUrl() . $file->path;
+        }
+
+        return null;
+    }
+
+    protected function getFileModel(string $uuid): FilesModel|null
     {
         return $this->framework
             ->getAdapter(FilesModel::class)
             ->findByUuid($uuid);
     }
 
-    protected function replaceInsertTags(string $content): string
+    protected function replaceInsertTags(string $value): string
     {
-        $controller = $this->framework->getAdapter(Controller::class);
+        return $this->insertTagParser->replaceInline($value);
+    }
 
-        $content = $controller->__call('replaceInsertTags', [$content, false]);
-        $content = $controller->__call('replaceInsertTags', [$content, true]);
+    protected function getCanonicalUrlForRequest(): string|null
+    {
+        $responseContext = $this->responseContextAccessor->getResponseContext();
 
-        return $content;
+        if (! $responseContext || ! $responseContext->has(HtmlHeadBag::class)) {
+            return null;
+        }
+
+        $request = $this->requestStack->getCurrentRequest();
+        if (! $request) {
+            return null;
+        }
+
+        return $responseContext->get(HtmlHeadBag::class)->getCanonicalUriForRequest($request);
     }
 }
