@@ -4,53 +4,44 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\Data\TwitterCards;
 
-use Contao\Model;
 use Hofff\Contao\SocialTags\Data\Data;
 use Hofff\Contao\SocialTags\Data\DataFactory;
-use Hofff\Contao\SocialTags\Data\Extractor;
+use Hofff\Contao\SocialTags\Data\ExtractorResolver;
 use Hofff\Contao\SocialTags\Data\Protocol;
 
 final class TwitterCardsFactory implements DataFactory
 {
-    /** @var Extractor */
-    private $extractor;
-
-    public function __construct(Extractor $extractor)
+    public function __construct(private readonly ExtractorResolver $resolver)
     {
-        $this->extractor = $extractor;
     }
 
-    public function generate(Model $reference, ?Model $fallback = null): Data
+    public function generate(object $reference, object|null $fallback = null): Data
     {
-        if (! $this->extractor->supports($reference, $fallback)) {
+        $extractor = $this->resolver->resolve(TwitterCardsExtractor::class, $reference, $fallback);
+        if (! $extractor instanceof TwitterCardsExtractor) {
             return new Protocol();
         }
 
-        $type = $fallback->hofff_st_twitter_type ?? null;
+        $type = $fallback?->hofff_st_twitter_type ?? null;
         if ($reference->hofff_st && $reference->hofff_st_twitter_type) {
             $type = $reference->hofff_st_twitter_type;
         }
 
-        switch ($type) {
-            case 'hofff_st_twitter_summary':
-                return new SummaryCardData(
-                    $this->extractor->extract('twitter', 'title', $reference, $fallback),
-                    $this->extractor->extract('twitter', 'site', $reference, $fallback),
-                    $this->extractor->extract('twitter', 'description', $reference, $fallback),
-                    $this->extractor->extract('twitter', 'image', $reference, $fallback)
-                );
-
-            case 'hofff_st_twitter_summary_large_image':
-                return new SummaryWithLargeImageCardData(
-                    $this->extractor->extract('twitter', 'title', $reference, $fallback),
-                    $this->extractor->extract('twitter', 'site', $reference, $fallback),
-                    $this->extractor->extract('twitter', 'description', $reference, $fallback),
-                    $this->extractor->extract('twitter', 'image', $reference, $fallback),
-                    $this->extractor->extract('twitter', 'creator', $reference, $fallback)
-                );
-
-            default:
-                return new Protocol();
-        }
+        return match ($type) {
+            'hofff_st_twitter_summary' => new SummaryCardData(
+                $extractor->extractTwitterTitle($reference, $fallback),
+                $extractor->extractTwitterSite($reference, $fallback),
+                $extractor->extractTwitterDescription($reference, $fallback),
+                $extractor->extractTwitterImage($reference, $fallback),
+            ),
+            'hofff_st_twitter_summary_large_image' => new SummaryWithLargeImageCardData(
+                $extractor->extractTwitterTitle($reference, $fallback),
+                $extractor->extractTwitterSite($reference, $fallback),
+                $extractor->extractTwitterDescription($reference, $fallback),
+                $extractor->extractTwitterImage($reference, $fallback),
+                $extractor->extractTwitterCreator($reference, $fallback),
+            ),
+            default => new Protocol(),
+        };
     }
 }
