@@ -4,38 +4,48 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\EventListener\Dca;
 
+use Contao\CoreBundle\DependencyInjection\Attribute\AsCallback;
 use Hofff\Contao\SocialTags\Util\TypeUtil;
 
 use function array_pad;
 use function explode;
-use function strpos;
+use function str_contains;
 
+#[AsCallback('tl_calendar_events', 'fields.hofff_st_og_type.options')]
+#[AsCallback('tl_faq', 'fields.hofff_st_og_type.options')]
+#[AsCallback('tl_news', 'fields.hofff_st_og_type.options')]
+#[AsCallback('tl_page', 'fields.hofff_st_og_type.options')]
 final class OpenGraphTypeOptions
 {
-    /** @var string[] */
-    private $types;
-
     /** @param string[] $types */
-    public function __construct(array $types)
+    public function __construct(private readonly array $types)
     {
-        $this->types = $types;
     }
 
-    /** @return string[] */
+    /** @return array<string, list<string>> */
     public function __invoke(): array
     {
         $options = [];
-        foreach ($this->types as $strType) {
-            if (strpos($strType, ' ') === false) {
-                [$strGroup, $strName]                                = array_pad(explode('.', $strType), 2, null);
-                TypeUtil::isStringWithContent($strName) || $strGroup = 'general';
-                $options[$strGroup][]                                = $strType;
+        $custom  = [];
+
+        foreach ($this->types as $type) {
+            if (! str_contains($type, ' ')) {
+                /** @psalm-var string $group */
+                [$group, $name] = array_pad(explode('.', $type), 2, null);
+
+                if (! TypeUtil::isStringWithContent($name)) {
+                    $group = 'general';
+                }
+
+                $options[$group][] = $type;
             } else {
-                $custom[] = $strType;
+                $custom[] = $type;
             }
         }
 
-        isset($custom) && $custom && $options['custom'] = $custom;
+        if ($custom !== []) {
+            $options['custom'] = $custom;
+        }
 
         return $options;
     }
