@@ -4,23 +4,20 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\EventListener\Hook;
 
+use Contao\CoreBundle\DependencyInjection\Attribute\AsHook;
 use Contao\PageModel;
 use Hofff\Contao\SocialTags\Data\SocialTagsFactory;
 use Symfony\Component\HttpFoundation\RequestStack;
 
+#[AsHook('generatePage')]
 final class PageSocialTagsListener extends SocialTagsDataAwareListener
 {
-    /** @var SocialTagsFactory */
-    private $factory;
-
-    public function __construct(RequestStack $requestStack, SocialTagsFactory $factory)
+    public function __construct(RequestStack $requestStack, private readonly SocialTagsFactory $factory)
     {
         parent::__construct($requestStack);
-
-        $this->factory = $factory;
     }
 
-    public function onGeneratePage(PageModel $page): void
+    public function __invoke(PageModel $page): void
     {
         if ($page->hofff_st === null || $this->getSocialTagsData()) {
             return;
