@@ -38,7 +38,7 @@ final class Plugin implements BundlePluginInterface, RoutingPluginInterface
     }
 
     /** @SuppressWarnings(PHPMD.UnusedFormalParameter) */
-    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel): ?RouteCollection
+    public function getRouteCollection(LoaderResolverInterface $resolver, KernelInterface $kernel): RouteCollection|null
     {
         $loader = $resolver->resolve(__DIR__ . '/../Resources/config/routing.xml');
         if (! $loader) {
