@@ -4,12 +4,10 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\Data;
 
-use Contao\Model;
-
 interface Extractor
 {
-    public function supports(Model $reference, ?Model $fallback = null): bool;
+    public function supports(object $reference, object|null $fallback = null): bool;
 
-    /** @return mixed */
-    public function extract(string $type, string $field, Model $reference, ?Model $fallback = null);
+    /** @return list<string> */
+    public function supportedDataContainers(): array;
 }
