@@ -4,34 +4,37 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\Data\Extractor;
 
-use Contao\File;
-use Contao\FilesModel;
-use Contao\Model;
 use Contao\News;
 use Contao\NewsModel;
 use Contao\PageModel;
 use Contao\StringUtil;
-use Hofff\Contao\SocialTags\Data\OpenGraph\OpenGraphImageData;
+use Hofff\Contao\SocialTags\Data\OpenGraph\OpenGraphExtractor;
+use Hofff\Contao\SocialTags\Data\OpenGraph\OpenGraphExtractorPlugin;
 use Hofff\Contao\SocialTags\Data\OpenGraph\OpenGraphType;
+use Hofff\Contao\SocialTags\Data\TwitterCards\TwitterCardsExtractor;
+use Hofff\Contao\SocialTags\Data\TwitterCards\TwitterCardsExtractorPlugin;
 use Hofff\Contao\SocialTags\Util\TypeUtil;
 
-use function array_pad;
-use function explode;
-use function is_file;
-use function method_exists;
 use function str_replace;
 use function strip_tags;
 use function stripos;
 use function trim;
-use function ucfirst;
 
 /**
+ * @implements OpenGraphExtractor<NewsModel, PageModel>
+ * @implements TwitterCardsExtractor<NewsModel, PageModel>
+ * @SuppressWarnings(PHPMD.TooManyPublicMethods)
  * @SuppressWarnings(PHPMD.UnusedPrivateMethod)
  * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
  */
-final class NewsExtractor extends AbstractExtractor
+final class NewsExtractor extends AbstractExtractor implements OpenGraphExtractor, TwitterCardsExtractor
 {
-    public function supports(Model $reference, ?Model $fallback = null): bool
+    /** @use OpenGraphExtractorPlugin<NewsModel, PageModel> */
+    use OpenGraphExtractorPlugin;
+    /** @use TwitterCardsExtractorPlugin<NewsModel, PageModel> */
+    use TwitterCardsExtractorPlugin;
+
+    public function supports(object $reference, object|null $fallback = null): bool
     {
         if (! $reference instanceof NewsModel) {
             return false;
@@ -40,100 +43,46 @@ final class NewsExtractor extends AbstractExtractor
         return $fallback instanceof PageModel;
     }
 
-    /** @return mixed */
-    public function extract(string $type, string $field, Model $reference, ?Model $fallback = null)
+    /** {@inheritDoc} */
+    public function supportedDataContainers(): array
     {
-        $methodName = 'extract' . ucfirst($type) . ucfirst($field);
-
-        if ($methodName !== __FUNCTION__ && method_exists($this, $methodName)) {
-            return $this->$methodName($reference, $fallback);
-        }
-
-        return null;
-    }
-
-    private function extractTwitterTitle(NewsModel $newsModel): ?string
-    {
-        if ($newsModel->hofff_st && TypeUtil::isStringWithContent($newsModel->hofff_st_twitter_title)) {
-            return $this->replaceInsertTags($newsModel->hofff_st_twitter_title);
-        }
-
-        return $this->getNewsTitle($newsModel);
-    }
-
-    private function extractTwitterSite(NewsModel $newsModel, PageModel $referencePage): ?string
-    {
-        if ($newsModel->hofff_st && $newsModel->hofff_st_twitter_site) {
-            return $newsModel->hofff_st_twitter_site;
-        }
-
-        return $referencePage->hofff_st_twitter_site ?: null;
-    }
-
-    private function extractTwitterDescription(NewsModel $newsModel): ?string
-    {
-        if ($newsModel->hofff_st && TypeUtil::isStringWithContent($newsModel->hofff_st_twitter_description)) {
-            return $this->replaceInsertTags($newsModel->hofff_st_twitter_description);
-        }
-
-        return $this->getNewsDescription($newsModel) ?: null;
-    }
-
-    private function extractTwitterImage(NewsModel $newsModel, PageModel $referencePage): ?string
-    {
-        $file = $this->getImage('hofff_st_twitter_image', $newsModel, $referencePage);
-
-        if ($file && is_file($this->projectDir . '/' . $file->path)) {
-            return $this->getBaseUrl() . $file->path;
-        }
-
-        return null;
-    }
-
-    private function extractTwitterCreator(NewsModel $newsModel, PageModel $referencePage): ?string
-    {
-        if ($newsModel->hofff_st && $newsModel->hofff_st_twitter_creator) {
-            return $newsModel->hofff_st_twitter_creator;
-        }
-
-        return $referencePage->hofff_st_twitter_creator ?: null;
+        return ['tl_news'];
     }
 
     /**
-     * @param string|resource $strImage
+     * Returns the meta description if present, otherwise the shortened teaser.
      */
-    private function extractOpenGraphImageData(NewsModel $newsModel, PageModel $referencePage): OpenGraphImageData
+    protected function getContentDescription(object $reference): string|null
     {
-        $imageData = new OpenGraphImageData();
-        $file      = $this->getImage('hofff_st_og_image', $newsModel, $referencePage);
-
-        if ($file && is_file($this->projectDir . '/' . $file->path)) {
-            $objImage = new File($file->path);
-            $imageData->setURL($this->getBaseUrl() . $file->path);
-            $imageData->setMIMEType($objImage->mime);
-            $imageData->setWidth($objImage->width);
-            $imageData->setHeight($objImage->height);
+        if (TypeUtil::isStringWithContent($reference->description)) {
+            return $this->replaceInsertTags(trim(str_replace(["\n", "\r"], [' ', ''], $reference->description)));
         }
 
-        return $imageData;
+        if (! TypeUtil::isStringWithContent($reference->teaser)) {
+            return null;
+        }
+
+        // Generate the description from the teaser the same way as the news reader does
+        $description = $this->replaceInsertTags($reference->teaser);
+        $description = strip_tags($description);
+        $description = str_replace("\n", ' ', $description);
+
+        return StringUtil::substr($description, 320);
     }
 
-    private function extractOpenGraphTitle(NewsModel $newsModel): ?string
+    protected function getContentTitle(object $reference): string
     {
-        if ($newsModel->hofff_st && TypeUtil::isStringWithContent($newsModel->hofff_st_og_title)) {
-            return $this->replaceInsertTags($newsModel->hofff_st_og_title);
-        }
-
-        return $this->getNewsTitle($newsModel) ?: null;
+        return (string) ($reference->pageTitle ?: $reference->headline);
     }
 
-    private function extractOpenGraphUrl(NewsModel $newsModel): string
+    protected function defaultOpenGraphType(): OpenGraphType
     {
-        if ($newsModel->hofff_st && TypeUtil::isStringWithContent($newsModel->hofff_st_og_url)) {
-            return $this->replaceInsertTags($newsModel->hofff_st_og_url);
-        }
+        return new OpenGraphType('article');
+    }
 
-        $newsUrl = News::generateNewsUrl($newsModel, false, true);
+    protected function getContentUrl(object $reference): string
+    {
+        $newsUrl = News::generateNewsUrl($reference, false, true);
 
         // Prepend scheme and host if URL is not absolute
         if (stripos($newsUrl, 'http') !== 0) {
@@ -141,94 +90,5 @@ final class NewsExtractor extends AbstractExtractor
         }
 
         return $newsUrl;
-    }
-
-    private function extractOpenGraphDescription(NewsModel $newsModel): ?string
-    {
-        if ($newsModel->hofff_st && TypeUtil::isStringWithContent($newsModel->hofff_st_og_description)) {
-            return $this->replaceInsertTags($newsModel->hofff_st_og_description);
-        }
-
-        return $this->getNewsDescription($newsModel) ?: null;
-    }
-
-    private function extractOpenGraphSiteName(NewsModel $newsModel, PageModel $fallback): string
-    {
-        if ($newsModel->hofff_st && TypeUtil::isStringWithContent($newsModel->hofff_st_og_site)) {
-            return $this->replaceInsertTags($newsModel->hofff_st_og_site);
-        }
-
-        return strip_tags($fallback->rootTitle);
-    }
-
-    private function extractOpenGraphType(NewsModel $newsModel): OpenGraphType
-    {
-        if ($newsModel->hofff_st && TypeUtil::isStringWithContent($newsModel->hofff_st_og_type)) {
-            [$namespace, $type] = array_pad(explode(' ', $newsModel->hofff_st_og_type, 2), 2, null);
-
-            if ($type === null) {
-                return new OpenGraphType($namespace);
-            }
-
-            return new OpenGraphType($type, $namespace);
-        }
-
-        return new OpenGraphType('article');
-    }
-
-    /**
-     * Returns the meta description if present, otherwise the shortened teaser.
-     */
-    private function getNewsDescription(NewsModel $model): ?string
-    {
-        if (TypeUtil::isStringWithContent($model->description)) {
-            return $this->replaceInsertTags(trim(str_replace(["\n", "\r"], [' ', ''], $model->description)));
-        }
-
-        if (! TypeUtil::isStringWithContent($model->teaser)) {
-            return null;
-        }
-
-        // Generate the description from the teaser the same way as the news reader does
-        $description = $this->replaceInsertTags($model->teaser ?? '', false);
-        $description = strip_tags($description);
-        $description = str_replace("\n", ' ', $description);
-        $description = StringUtil::substr($description, 320);
-
-        return $description;
-    }
-
-    /**
-     * Returns the meta title if present, otherwise the headline.
-     */
-    private function getNewsTitle(NewsModel $model): ?string
-    {
-        $title = $model->pageTitle ?: $model->headline;
-        if (TypeUtil::isStringWithContent($title)) {
-            return $this->replaceInsertTags($title);
-        }
-
-        return null;
-    }
-
-    /**
-     * Retrieves an image from the news for a given key. It fallbacks to the news image or page image if not defined.
-     */
-    private function getImage(string $key, NewsModel $newsModel, PageModel $referencePage): ?FilesModel
-    {
-        $image = null;
-        if ($newsModel->hofff_st && $newsModel->{$key}) {
-            $image = $newsModel->{$key};
-        } elseif ($newsModel->addImage && $newsModel->singleSRC) {
-            $image = $newsModel->singleSRC;
-        } elseif ($referencePage->{$key}) {
-            $image = $referencePage->{$key};
-        } else {
-            return null;
-        }
-
-        return $this->framework
-            ->getAdapter(FilesModel::class)
-            ->findByUuid($image);
     }
 }
