@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\DependencyInjection;
 
+use Hofff\Contao\SocialTags\Data\DataFactory;
+use Hofff\Contao\SocialTags\Data\Extractor;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
@@ -12,7 +14,7 @@ use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 final class HofffContaoSocialTagsExtension extends Extension
 {
     /**
-     * @param mixed[][] $configs
+     * {@inheritDoc}
      *
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
@@ -20,11 +22,14 @@ final class HofffContaoSocialTagsExtension extends Extension
     {
         $loader = new XmlFileLoader(
             $container,
-            new FileLocator(__DIR__ . '/../Resources/config')
+            new FileLocator(__DIR__ . '/../Resources/config'),
         );
 
         $loader->load('config.xml');
         $loader->load('services.xml');
         $loader->load('listeners.xml');
+
+        $container->registerForAutoconfiguration(Extractor::class)->addTag(Extractor::class);
+        $container->registerForAutoconfiguration(DataFactory::class)->addTag(DataFactory::class);
     }
 }
