@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\EventListener\Hook;
 
-use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
+use Contao\CoreBundle\DependencyInjection\Attribute\AsHook;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Contao\CoreBundle\Routing\ScopeMatcher;
 use Contao\Input;
 use Contao\Model;
@@ -14,34 +15,18 @@ use Contao\StringUtil;
 use Hofff\Contao\SocialTags\Data\SocialTagsFactory;
 use Symfony\Component\HttpFoundation\RequestStack;
 
-use function version_compare;
-
-use const VERSION;
-
 final class NewsReaderListener extends SocialTagsDataAwareListener
 {
-    /** @var SocialTagsFactory */
-    private $factory;
-
-    /** @var ScopeMatcher */
-    private $scopeMatcher;
-
-    /** @var ContaoFrameworkInterface */
-    private $framework;
-
     public function __construct(
         RequestStack $requestStack,
-        SocialTagsFactory $factory,
-        ScopeMatcher $scopeMatcher,
-        ContaoFrameworkInterface $framework
+        private readonly SocialTagsFactory $factory,
+        private readonly ScopeMatcher $scopeMatcher,
+        private readonly ContaoFramework $framework,
     ) {
         parent::__construct($requestStack);
-
-        $this->factory      = $factory;
-        $this->scopeMatcher = $scopeMatcher;
-        $this->framework    = $framework;
     }
 
+    #[AsHook('getContentElement')]
     public function onGetContentElement(Model $model, string $result): string
     {
         if ($model->type !== 'module') {
@@ -56,9 +41,10 @@ final class NewsReaderListener extends SocialTagsDataAwareListener
         return $this->onGetFrontendModule($module, $result);
     }
 
+    #[AsHook('getFrontendModule')]
     public function onGetFrontendModule(ModuleModel $model, string $result): string
     {
-        $request = $this->requestStack->getMasterRequest();
+        $request = $this->requestStack->getMainRequest();
         if (! $request || ! $this->scopeMatcher->isFrontendRequest($request)) {
             return $result;
         }
@@ -71,7 +57,7 @@ final class NewsReaderListener extends SocialTagsDataAwareListener
 
         $newsModel = $this->getNewsModel($model);
         if ($newsModel) {
-            $this->setSocialTagsData($this->factory->generateByModel($newsModel));
+            $this->setSocialTagsData($this->factory->generate($newsModel));
         }
 
         return $result;
@@ -82,20 +68,22 @@ final class NewsReaderListener extends SocialTagsDataAwareListener
         return $model->type === 'newsreader';
     }
 
-    private function getNewsModel(ModuleModel $model): ?NewsModel
+    private function getNewsModel(ModuleModel $model): NewsModel|null
     {
+        /** @psalm-suppress RiskyTruthyFalsyComparison */
         return NewsModel::findPublishedByParentAndIdOrAlias(
-            $this->framework->getAdapter(Input::class)->get('items'),
-            StringUtil::deserialize($model->news_archives, true)
+            $this->framework->getAdapter(Input::class)->get('auto_item'),
+            StringUtil::deserialize($model->news_archives, true),
         );
     }
 
     private function determineModuleModel(ModuleModel $model): ModuleModel
     {
+        /** @psalm-suppress RiskyTruthyFalsyComparison */
         if (
-            ($model->type === 'newsarchive' || (version_compare(VERSION, '4.7', '>=') && $model->type === 'newslist'))
+            ($model->type === 'newsarchive' || $model->type === 'newslist')
             && $model->news_readerModule > 0
-            && $this->framework->getAdapter(Input::class)->get('items')
+            && $this->framework->getAdapter(Input::class)->get('auto_item')
         ) {
             $readerModel = ModuleModel::findByPk($model->news_readerModule);
             if ($readerModel) {
