@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace Hofff\Contao\SocialTags\EventListener\Hook;
 
+use Contao\CoreBundle\DependencyInjection\Attribute\AsHook;
 use Hofff\Contao\SocialTags\Data\Data;
 
+#[AsHook('generatePage', priority: -1)]
 final class SocialTagsDataInjectionListener extends SocialTagsDataAwareListener
 {
     /** @SuppressWarnings(PHPMD.Superglobals) */
-    public function onGeneratePage(): void
+    public function __invoke(): void
     {
         $socialTagsData = $this->getSocialTagsData();
         if (! $socialTagsData instanceof Data) {
