@@ -2,16 +2,10 @@
 
 declare(strict_types=1);
 
-use Hofff\Contao\SocialTags\EventListener\Dca\OpenGraphTypeOptions;
-use Hofff\Contao\SocialTags\EventListener\Dca\PageDcaListener;
-
-$this->loadLanguageFile('hofff_st');
-
 $GLOBALS['TL_DCA']['tl_page']['list']['operations']['hofff_st_og_facebookLint'] = [
     'label'      => &$GLOBALS['TL_LANG']['tl_page']['hofff_st_og_facebookLint'],
     'icon'       => 'bundles/hofffcontaosocialtags/images/og.png',
     'attributes' => ' onclick="window.open(this.href); return false;"',
-    'button_callback' => [PageDcaListener::class, 'facebookLinkButton'],
 ];
 
 $GLOBALS['TL_DCA']['tl_page']['palettes']['__selector__'][] = 'hofff_st';
@@ -25,7 +19,7 @@ foreach ($GLOBALS['TL_DCA']['tl_page']['palettes'] as $strKey => &$strPalette) {
     $strPalette = preg_replace(
         '@(\{meta_legend\}[^;]*;)@',
         '$1{hofff_st_legend},hofff_st;',
-        $strPalette
+        $strPalette,
     );
 }
 
@@ -78,7 +72,6 @@ $GLOBALS['TL_DCA']['tl_page']['fields']['hofff_st_og_type'] = [
     'exclude'   => true,
     'inputType' => 'select',
     'default'   => 'website',
-    'options_callback'   => [OpenGraphTypeOptions::class, '__invoke'],
     'reference' => &$GLOBALS['TL_LANG']['hofff_st']['types'],
     'eval'      => [
         'mandatory'      => true,
